перенесла рендер функцию в отдельный модуль
