+++ hwDraftComments_Модульный_JS/renderCommentators.js
@@ -1,0 +1,32 @@
+
+const listElement = document.getElementById ('list');
+
+export const renderCommentators = (сommentators, initEventListeners, editEventListeners, answerComment ) => {
+  const commentatorsHtml = сommentators.map((сommentator, index) => {
+    return `<li class="comment" data-index="${index}">
+      <div class="comment-header">
+       <div>${сommentator.name}</div>
+       <div>${сommentator.date}</div>
+      </div>
+      <div class="comment-body">
+       <div class="${сommentator.isEdit ? 'display-none' : 'comment-text'}">${сommentator.comment}</div>
+       <textarea type="textarea" class="${сommentator.isEdit ? 'add-form-text' : 'display-none'}" >${сommentator.comment}</textarea>
+      </div>
+      <div class="comment-footer">
+        <button class="${сommentator.isEdit ? 'edit-button' : 'save-button'}" data-index="${index}">Редактировать</button>
+        <button class="${сommentator.isEdit ? 'save-button' : 'textarea'}" data-index="${index}">Сохранить</button>
+      <div class="likes">
+        <span class="likes-counter">${сommentator.likes}</span>
+        <button data-like="${сommentator.likes}" data-index="${index}" class="like-button ${сommentators[index].isLiked ?'-active-like' : 'like-button'}"></button>
+      </div>
+    </div>
+    </li>`;
+  })
+  .join("");
+
+  listElement.innerHTML = commentatorsHtml;
+  
+  initEventListeners();
+  editEventListeners();
+  answerComment();
+};
